Add README file for JavaScript 30-day learning plan
This commit adds a new README file to the repository that outlines a detailed 30-day plan for learning JavaScript. The plan includes a variety of resources, such as video tutorials, interactive exercises, and recommended readings, to help learners achieve mastery of the language.

The commit includes the following changes:

- Created a new file called "README.md"
- Added an introduction to the purpose and benefits of the learning plan
- Listed the topics that will be covered over the course of 30 days, with links to resources for each topic
- Added details about the structure and format of the learning plan, including recommendations for pacing and flexibility
- Added a final section on continuing education and setting goals for the future

The README file will be a valuable resource for anyone looking to learn JavaScript in a structured and comprehensive way.